tutorialProject/requirements.txt
Django==1.10
django-filters==0.2.1
django-tinymce==2.3.0
pip==8.1.2
psycopg2==2.6.2
pytz==2016.6.1
django-registration==2.1.2
Pillow==3.3.1
django-countries==4.0
coverage==4.2
ipython==4.0.0
ipdb==0.10.1
django-modalview==0.1.5
green
requests==2.11.1
pypng==0.0.18
django-debug-toolbar
django-material
django-cachalot
python-memcached
